Use partial points for ranking show points scored in the scoreboard, even when the solution is not entirely correct

+++ webapp/src/Utils/Scoreboard/TeamScore.php
@@ -6,7 +6,7 @@ use App\Entity\Team;
 
 class TeamScore
 {
-    public int $numPoints = 0;
+    public float $numPoints = 0;
 
     /** @var float[] */
     public array $solveTimes = [];
